backend/pyproject.toml
[project]
name = "news-intelligence-backend"
version = "0.1.0"
description = "AI-powered news intelligence platform backend"
authors = [
    {name = "Your Name", email = "you@example.com"},
]
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    "pydantic-settings>=2.1.0",
    "redis>=5.0.1",
    "celery>=5.3.6",
    "python-dotenv>=1.0.1",
    "httpx>=0.26.0",
]
requires-python = ">=3.11"
